2.5.506+ 标准时间小数化

+++ rvs2.0/src/com/osh/rvs/action/inline/PositionPanelAction.java
@@ -1146,7 +1146,7 @@ public class PositionPanelAction extends BaseAction {
 				/// 取得本次工时
 				String sUse_seconds = RvsUtils.getZeroOverLine("_default", null, user, user.getProcess_code());
 				try {
-					use_seconds = Integer.parseInt(sUse_seconds) * 60;
+					use_seconds = (int) (Double.parseDouble(sUse_seconds) * 60);
 				} catch (Exception e){
 				}
